requirements.txt
google-cloud-compute
google-cloud-artifact-registry
python-dotenv